dumroo-admin/requirements.txt
streamlit==1.38.0
pandas==2.2.2
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
python-dotenv==1.0.1
pydantic==2.8.2